provider/aws: move to oldaws
