+++ go/src/koding/kites/kloud/provider/oldaws/bootstrap.go
@@ -1,4 +1,4 @@
-package awsprovider
+package oldaws
 
 import (
 	"bytes"
