+++ go/src/koding/kites/kloud/provider/oldaws/info.go
@@ -1,4 +1,4 @@
-package awsprovider
+package oldaws
 
 import (
 	"time"
