+++ go/src/koding/kites/kloud/provider/oldaws/provider.go
@@ -1,4 +1,4 @@
-package awsprovider
+package oldaws
 
 import (
 	"fmt"
