+++ go/src/koding/kites/kloud/provider/oldaws/stackplan.go
@@ -1,4 +1,4 @@
-package awsprovider
+package oldaws
 
 import (
 	"fmt"
