+++ go/src/koding/kites/kloud/provider/oldaws/start.go
@@ -1,4 +1,4 @@
-package awsprovider
+package oldaws
 
 import (
 	"errors"
