+++ go/src/koding/kites/kloud/queue/queue.go
@@ -6,14 +6,14 @@ import (
 	"github.com/koding/logging"
 
 	"koding/kites/kloud/machinestate"
-	"koding/kites/kloud/provider/aws"
+	"koding/kites/kloud/provider/oldaws"
 
 	"gopkg.in/mgo.v2"
 	"gopkg.in/mgo.v2/bson"
 )
 
 type Queue struct {
-	AwsProvider *awsprovider.Provider
+	AwsProvider *oldaws.Provider
 	Log         logging.Logger
 }
