+++ go/src/koding/kites/kloud/stackplan/template_test.go
@@ -11,7 +11,7 @@ import (
 	"github.com/koding/logging"
 
 	"koding/kites/kloud/provider"
-	"koding/kites/kloud/provider/aws"
+	"koding/kites/kloud/provider/oldaws"
 	"koding/kites/kloud/stackplan"
 )
 
@@ -384,7 +384,7 @@ func TestTerraformTemplate_SetAWSRegion(t *testing.T) {
 	}
 
 	// TODO(rjeczalik): move the test to provider/aws
-	s := &awsprovider.Stack{BaseStack: &provider.BaseStack{Builder: &stackplan.Builder{Template: template}}}
+	s := &oldaws.Stack{BaseStack: &provider.BaseStack{Builder: &stackplan.Builder{Template: template}}}
 	if err := s.SetAwsRegion("us-east-1"); err != nil {
 		t.Fatal(err)
 	}
